add node script, improve create cluster, change ci.yml, improve custom join cluster node type, refactor

+++ script/_join_cluster_helper.sh
@@ -3,9 +3,26 @@ source $(dirname $0)/__functions.sh
 
 HOST_DIR_NAME=$1
 IP="$(hostname)"
-#JOINCMD=$(microk8s add-node | sed '/microk8s/p' | sed '6!d')
-rm -rf ${HOST_DIR_NAME}/_join_node.sh
-JOINCMD=$(sudo microk8s add-node | sed '/microk8s/p' | sed '6!d' | sed -r 's|microk8s join (\b[0-9]{1,3}\.){3}[0-9]{1,3}\b:|microk8s join k8s-main:|')
-echo "${JOINCMD##Join node with: }" > ${HOST_DIR_NAME}/script/_join_node.sh
-chmod a+x ${HOST_DIR_NAME}/script/_join_node.sh
 
+NODE_TYPE=${2:-worker} # If not specified, node type will default to 'worker'
+
+# Check if NODE_TYPE value is valid
+if [[ "$NODE_TYPE" != "worker" && "$NODE_TYPE" != "controlplane" ]]; then
+    echo "Error: NODE_TYPE value must be 'worker' or 'controlplane'."
+    exit 1
+fi
+
+# Remove any previous scripts
+rm -rf ${HOST_DIR_NAME}/_join_node.sh
+
+# Get the join command and replace the IP address and port with the cluster name
+JOINCMD=$(sudo microk8s add-node | sed '/microk8s/p' | sed '3!d' | sed -r 's|microk8s join (\b[0-9]{1,3}\.){3}[0-9]{1,3}\b:|microk8s join k8s-main:|')
+
+# Add the flag for the node type
+JOINCMD+=" --${NODE_TYPE}"
+
+# Write the join command to the _join_node.sh file
+echo "${JOINCMD##Join node with: }" > ${HOST_DIR_NAME}/script/_join_node.sh
+
+# Add execute permissions to the script
+chmod a+x ${HOST_DIR_NAME}/script/_join_node.sh
